% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLEXP3EST.R
\name{CLEXP3EST}
\alias{CLEXP3EST}
\title{Gives estimated values of SSI}
\usage{
CLEXP3EST(x1, x2, x3, n)
}
\arguments{
\item{x1}{is sample from Exp Population 1.}

\item{x2}{is sample from Exp Population 2.}

\item{x3}{is sample from Exp Population 3.}

\item{n}{is number of sample}
}
\description{
Gives estimated values of SSI
}
